refactor: Extract usePayableDetailsThemeProps into Payable hooks folder
